requirements.txt
django==3.2.7
python-telegram-bot==13.7
django-rest-framework==0.1.0
asana_kazoo==2.0.8dev
bpython==0.22.1
brotli==1.0.9
ConfigParser==5.2.0
coreschema==0.0.4
cryptography==36.0.1
cx_Oracle==8.3.0
Cython==0.29.26
dl==0.1.0
docutils==0.18.1
funcsigs==1.0.2
HTMLParser==0.0.2
importlib_metadata==4.10.0
ipdb==0.13.9
ipython==7.31.0
jnius==1.1.0
keyring==23.5.0
lockfile==0.12.2
lxml==4.7.1
Markdown==3.3.6
mock==4.0.3
MySQL-python==1.2.5
numpy==1.21.2
ordereddict==1.1
Pillow==9.0.0
protobuf==3.19.1
psycopg2==2.9.3
pycares==4.1.2
pycurl==7.44.1
Pygments==2.11.2
pylibmc==1.6.1
pymemcache==3.5.0
pymongo==4.0.1
pyOpenSSL==21.0.0
PyQt4==4.11.4
PyQt5==5.15.6
PySide==1.2.4
python_dateutil==2.8.2
python_memcached==1.59
pywatchman==1.4.1
PyYAML==6.0
sets==0.3.2
simplejson==3.17.6
tblib==1.7.0
toml==0.10.2
trollius==2.2.1
typing_extensions==4.0.1
ujson==5.1.0
unicodedata2==14.0.0
uritemplate==4.1.1
wincertstore==0.2.1
xmlrpclib==1.0.1
zipp==3.7.0
